requirements.txt
# Core
streamlit>=1.30.0
pandas>=2.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Google Gemini AI SDK
google-generativeai>=0.1.0  # adjust version if needed

# Optional / utility
pytest>=7.0.0      # for running tests

